+++ CHANGELOG.md
@@ -1,9 +1,13 @@
+## 0.5.3
+Added two new methods "onHoverInBarrier" and "dismissCondition" to the 'PopupController.show' 
+method. 
+
 ## 0.5.2
-Now, on setting [barrierDismissible] to 'true', will also cause 'IgnorePointer', on the barrier.
-Note:- The original pop-up [child] will still able to detect all mouse events.
+Now, on setting "barrierDismissible" to 'true', will also cause 'IgnorePointer', on the barrier.
+Note:- The original pop-up "child" will still able to detect all mouse events.
 
 ## 0.5.1
-Added [onDismiss] parameter for 'PopupController'. Can set this parameter from either 
+Added "onDismiss" parameter for 'PopupController'. Can set this parameter from either 
 'PopupController.show' method, or using 'PopupController.of' static method.
 
 ## 0.5.0+1